+++ src/main/java/umc/th/juinjang/controller/ImageController.java
@@ -1,43 +1,28 @@
 package umc.th.juinjang.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
-import org.springframework.http.StreamingHttpOutputMessage.Body;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.Mapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import umc.th.juinjang.apiPayload.ApiResponse;
 import umc.th.juinjang.apiPayload.code.status.SuccessStatus;
-import umc.th.juinjang.converter.limjang.LimjangPostConverter;
 import umc.th.juinjang.model.dto.image.ImageDeleteRequestDTO;
 import umc.th.juinjang.model.dto.image.ImageListResponseDTO;
-import umc.th.juinjang.model.dto.image.ImageUploadRequestDTO;
-import umc.th.juinjang.model.dto.image.ImageUploadRequestDTO.ImageDto;
-import umc.th.juinjang.model.dto.image.ImageUploadResponseDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangPostRequestDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangPostResponseDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangPostResponseDTO.PostDTO;
 import umc.th.juinjang.service.ImageService.ImageCommandService;
 import umc.th.juinjang.service.ImageService.ImageQueryService;
 import umc.th.juinjang.service.LimjangService.LimjangCommandService;
-import umc.th.juinjang.service.LimjangService.LimjangQueryService;
 
 @RestController
 @RequestMapping("/api/limjang/image")
