+++ src/main/java/umc/th/juinjang/converter/limjang/LimjangMainListConverter.java
@@ -5,7 +5,7 @@ import static umc.th.juinjang.service.LimjangService.LimjangPriceBridge.makePric
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
-import umc.th.juinjang.model.dto.limjang.LimjangMainViewListResponsetDTO;
+import umc.th.juinjang.model.dto.limjang.response.LimjangMainViewListResponsetDTO;
 import umc.th.juinjang.model.entity.Image;
 import umc.th.juinjang.model.entity.Limjang;
 import umc.th.juinjang.model.entity.LimjangPrice;
