+++ src/main/java/umc/th/juinjang/model/dto/limjang/request/LimjangUpdateRequestDTO.java
@@ -1,4 +1,4 @@
-package umc.th.juinjang.model.dto.limjang;
+package umc.th.juinjang.model.dto.limjang.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
