+++ src/main/java/umc/th/juinjang/model/dto/limjang/response/LimjangMemoResponseDTO.java
@@ -1,4 +1,4 @@
-package umc.th.juinjang.model.dto.limjang;
+package umc.th.juinjang.model.dto.limjang.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
