+++ src/main/java/umc/th/juinjang/model/dto/limjang/response/LimjangPostResponseDTO.java
@@ -1,4 +1,4 @@
-package umc.th.juinjang.model.dto.limjang;
+package umc.th.juinjang.model.dto.limjang.response;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
