+++ src/main/java/umc/th/juinjang/model/dto/limjang/response/LimjangTotalListResponseDTO.java
@@ -1,31 +1,34 @@
-package umc.th.juinjang.model.dto.limjang;
+package umc.th.juinjang.model.dto.limjang.response;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import umc.th.juinjang.model.dto.limjang.enums.LimjangCheckListVersionEnum;
 
-public class LimjangDetailResponseDTO {
+public class LimjangTotalListResponseDTO {
 
   @Builder
   @Getter
   @NoArgsConstructor
   @AllArgsConstructor
-  public static class DetailDto {
+  public static class TotalListDto {
+    private List<LimjangListDto> limjangList;
+  }
+
+  @Builder
+  @Getter
+  @NoArgsConstructor
+  @AllArgsConstructor
+  public static class LimjangListDto {
     private Long limjangId;
-    private LimjangCheckListVersionEnum checkListVersion;
     private List<String> images;
     private Integer purposeCode;
+    private Boolean isScraped;
     private String nickname;
     private Integer priceType;
     private List<String> priceList;
+    private String totalAverage;
     private String address;
-    private String addressDetail;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
   }
-
 }
