+++ src/main/java/umc/th/juinjang/service/LimjangService/LimjangCommandService.java
@@ -1,10 +1,8 @@
 package umc.th.juinjang.service.LimjangService;
 
-import java.util.List;
-import umc.th.juinjang.model.dto.limjang.LimjangDeleteRequestDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangPostRequestDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangUpdateRequestDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangUpdateRequestDTO.UpdateDto;
+import umc.th.juinjang.model.dto.limjang.request.LimjangDeleteRequestDTO;
+import umc.th.juinjang.model.dto.limjang.request.LimjangPostRequestDTO;
+import umc.th.juinjang.model.dto.limjang.request.LimjangUpdateRequestDTO;
 import umc.th.juinjang.model.entity.Limjang;
 import umc.th.juinjang.model.entity.Member;
