+++ src/main/java/umc/th/juinjang/service/LimjangService/LimjangCommandServiceImpl.java
@@ -15,9 +15,9 @@ import umc.th.juinjang.apiPayload.exception.handler.LimjangHandler;
 import umc.th.juinjang.apiPayload.exception.handler.MemberHandler;
 import umc.th.juinjang.converter.limjang.LimjangPostConverter;
 import umc.th.juinjang.converter.limjang.LimjangUpdateConverter;
-import umc.th.juinjang.model.dto.limjang.LimjangDeleteRequestDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangPostRequestDTO.PostDto;
-import umc.th.juinjang.model.dto.limjang.LimjangUpdateRequestDTO;
+import umc.th.juinjang.model.dto.limjang.request.LimjangDeleteRequestDTO;
+import umc.th.juinjang.model.dto.limjang.request.LimjangPostRequestDTO.PostDto;
+import umc.th.juinjang.model.dto.limjang.request.LimjangUpdateRequestDTO;
 import umc.th.juinjang.model.entity.Limjang;
 import umc.th.juinjang.model.entity.LimjangPrice;
 import umc.th.juinjang.model.entity.Member;
